Make `MachinePistol` and work on bullets
Also modified the way weapons work. All weapons and bullet templates are scenes that share a script. In order to keep it as simple as possible, each scene has a prescribed layout they need in order to work plus a unique statistics resource for all of their unique information. Look at `DefaultBullet` and `MachinePistol` for examples