Change hitbox component to hit detector component
It is an Area2D that copies all of the character's hitboxes for itself and syncs collision layers and masks with the character. It could be very expensive to copy hitboxes like this if a character has more than one and there are a lot of characters.

However, characters should only have one and I don't expect to have too many characters and this method simply works well. Premature optimization is blah blah blah