Implement bullet hit detection and damage dealing
The life component has a lot of layers of abstraction in it, but it's not hard to understand, shouldn't impact performance, and might be useful later. It's fine.

I'm struggling to think of how to pass along the attacker's identity so the game can announce who killed whom and easily track damage dealt statistics by character and what have you. I will probably need to couple bullet to characters a little.