Modify all character-related signals to pass along character ID; format some code
I realized that when more characters are added, all of these events routed through the event bus would be listened to by everyone; they need to be distinguished by ID. Now I need to ensure when callbacks are performed by character's components, they first check if the id's match. I think I got them all for now.

This doesn't seem terribly inefficient to me. I think it's better than finding other ways for all of these callbacks to communicate so consistently and dynamically and simply. If it becomes a performance issue, I'll cross that bridge when I come to it.