Implement item slot cycling
It just works after about an hour of effort. You can add any number of items to `item_inventory` and it will just work. That array is populated by all of the types in `Constants.item_types` so it should just work forever.

Note the weird way you have to translate enum values and strings back and forth because GDScript doesn't have string enums or enum constants (besides ints). I commented the item slot cycling functions thoroughly because of this. I've actually got the hang of it now.