Implement object pooling for bullets
It might behoove me to factor this out of Weapon and into its own script(s), but it works well for now. I really need to implement a way to pool just enough or to simply stop getting the next one when there isn't one that isn't active.

Bullet is a little more complicated now, but I don't think it will be too much to work with. Consider how much mental effort is saved by having one script for all Bullet scenes when thinking of the extra mental effort to understand Bullet. The only complexities are that now Bullets don't expire. They deactivate on ready, get activated when fired, then deactivate again.