Save first version of default.xlf
I believe this is correct XLIFF. Note that it doesn't have a target language. That is fine in XLIFF, and this particular file will only have all of the English strings. Other documents for each languge will be added if translation work is ever actually done.

I intend to parse these files and make each `<file>` section into a .tres file in Godot, with each `<unit>` being a group of fields (which are parsed from `<segment>`s).

For example, this file will be parsed into a .tres with two sections, one for each weapon. Right now each section will just have a name field, but later descriptions could easily be added.

The idea is to get the user's language on first startup and build all of the string tables for their language, then use them for every future boot (unless the data is deleted). Every occurrence of end user visible human-readable strings in the game will reference this global data, which will have a different value based on language.

This work will allow flipping a switch to change the game's language (assuming the translation work is done). How about UI stuff? For this small game, I don't intend to put text in any graphics so I can solve that problem another day. :)