Modify default.xlf a bit and start work on XLIFF to Resource parser
I may have hit a wall. To create custom Resources with programmatically set exported vars, you have to build a GDScript class for each kind.

I would have to build a string for the .gd file for each <file>, create the file, then create Resources from that script and parse the <file>s again (or store all of the info in an array, actually) to set the values.

I think this is all doable in GDScript, so I'll figure it out. This probably would have been way easier in C#, but hey, it's a good learning experience.