Implement automated XLIFF to .tres conversion for localization strings
Workflow is to add a new culture XLIFF file to `assets/xliff/` with filename <culture-name>.xlf, then add a new <path> with the same culture-name as its contents to `xliff_paths.xml`. Add the translations to the <target> elements in the new .xlf file (there MUST be <target> elements for each <source> element) and your work is done.

Right now the code runs in the editor on startup and there's some XMLParser error I don't understand that doesn't seem to hurt anything. Ignore that. If asked about localization_string_table.gd, discard and reload to get the newly built version. I must move this code to an editor button or something soon. I must clean it up a lot, LATER. This works fine since it's probably just me working on this game for now. I'll add docs to the wiki about this workflow in a day or two.

Sometimes it also needs another reset to show the .tres files. You don't actually need to edit them since it's all automated, so that's not a big issue. They do need to be referenced for using their data, though (file and property names used anywhere scripts need the strings).

Next: writing the game's code to get these