Remove localization system, clean up some TODOs, and...
...associate `ItemEffect` with `Item` instead of other way around (effects shouldn't require any external information to work, build reusable effects that have properties set on use), clean up some documentation, and hardcode name strings.

I gutted the localization system because it's a rabbit hole with no point on this project. It was a good learning experience and I will learn more about building a proper XML localization system for my games going forward. I'm going to practice with MonoGame's XML workflow to get a feel for how to employ data-driven design with XML in all my games, even with Godot.